+++ settings/DevHome.Settings/Views/ExperimentalFeaturesPage.xaml
@@ -4,8 +4,9 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:behaviors="using:DevHome.Common.Behaviors"
+    xmlns:controls="using:CommunityToolkit.WinUI.Controls"
     xmlns:converters="using:CommunityToolkit.WinUI.Converters"
-    xmlns:vm="using:DevHome.Settings.ViewModels"
+    xmlns:settings="using:DevHome.Settings"
     behaviors:NavigationViewHeaderBehavior.HeaderMode="Never">
     <Page.Resources>
         <converters:DoubleToVisibilityConverter x:Key="CountToVisibilityConverter" GreaterThan="0" FalseValue="Visible" TrueValue="Collapsed"/>
@@ -28,14 +29,10 @@
             <StackPanel Orientation="Vertical">
                 <ItemsRepeater ItemsSource="{x:Bind ViewModel.Features}">
                     <ItemsRepeater.ItemTemplate>
-                        <DataTemplate x:DataType="vm:ExperimentalFeature">
-                            <Grid ColumnDefinitions="*,Auto"  Padding="4" Margin="4">
-                                <StackPanel Orientation="Vertical">
-                                    <TextBlock Text="{x:Bind Name, Mode=OneWay}" Margin="0,0,10,0" VerticalAlignment="Center"/>
-                                    <TextBlock Text="{x:Bind Description, Mode=OneWay}"/>
-                                </StackPanel>
+                        <DataTemplate x:DataType="settings:ExperimentalFeature">
+                            <controls:SettingsCard Header="{x:Bind Name, Mode=OneWay}" Description="{x:Bind Description, Mode=OneWay}" Margin="{ThemeResource SettingsCardMargin}">
                                 <ToggleSwitch IsOn="{x:Bind IsEnabled,Mode=TwoWay}" Grid.Column="1"/>
-                            </Grid>
+                            </controls:SettingsCard>
                         </DataTemplate>
                     </ItemsRepeater.ItemTemplate>
                 </ItemsRepeater>
